Consolidate rules, refiners, and configurations
This brings related units closer to each other, hopefully reducing the
risk of introducing cyclic dependencies.

The unit tests of rules have been converted to integration tests in
`Rule.tests` to test all rules at once in hope of revealing rules that
conflict with each other and to test commit refiners in conjunction
with all rules.