Disable Dependabot version updates
This is hopefully just a temporary solution.

Dependabot is not yet able to update multiple related dependencies
within a single commit.
For example, the versions of `@typescript-eslint/eslint-plugin` and
`@typescript-eslint/parser` must always match each other. The same is
true for `vitest`, `@vitest/coverage-c8`, and `@vitest/ui`.

Another issue is that Dependabot does not seem to de-duplicate
dependencies in the Yarn cache, thus leaving unused dependency artifacts
in the `.yarn/cache` directory.