Add token configuration for issue links
Issue link keys must be natural numbers with a string prefix. GitHub-
and GitLab-style issue links are supported by default.

Other tokens like squash markers and dependency versions are not
configurable for now.

This makes the tool easier for users to set up, as they won't have to
deal with regular expressions when configuring tokens. It also makes the
code simpler to maintain.